Noun and Verb Positions
Without having to go to the trouble of counting all the characters, I used the 'string.find' method to accurately determine the character location of the words "NOUN" and "VERB"
In this case, NOUN was at position 125 and VERB at position 171. Using 'string.find' is a great way to be able to locate text or values within a large data set.