String Manipulation & Slicing
The following code illustrates several instances of string manipulation by way of sub-sequencing. This was challenging for me because I had to physically count the number of characters in order to know the character value that I will use. Additionally, this includes a bit of string slicing and concatenation of the same.
It was a bit hard remembering where to place the (colon) : when subsequencing but I was able to crack it after several attempts.