String.Replace
In this example, I was able to use the string.replace() function to replace the terms "NOUN" and "VERB" with "duck" and "waddle" respectively. I can see how applicable the replace function could be when there is a large code base and certain words or texts have to be updated to reflect a change in the output. String.replace() becomes an efficient way to update or rectify information without having to sift through a lot of data and code