Function Examples
Here I learnt the general applications of using functions(aka procedures). Functions can be passed with arguments and called later on using 'return'.
When called through a print statement, the interpreter jumps back to the main body of the code and assigns the value of the print statement to the argument. E.g print rest_of_string("audacity") - the string audacity is assigned to the argument 's'. Thereafter, the interpreter jumps back to the return statement and executes the value of s[1:]