Defining and Using Functions
In this example, I was tasked with creating code that outputs the square of a given value. In this case, I defined a function 'square' with one argument 'a'. Thereafter, I returned the square of 'a' by introducing another variable 'c' = a * a. I returned c and printed out "print square (5)"
After learning this, I got introduced to function/procedure composition. This is where the output of one function is used as the input of another function. This is very common in programming and helps to save time and write efficient code.