Conditional Statements
If is a statement that runs or skips code based on whether a condition is true or false.  An if statement is followed by a condition that is expressed as a boolean and followed by a colon. Else is a statement that executes code when the condition is false. It does not require a boolean expression but is followed by a colon. Elif (short for else if) is a statement that executes code in situations where there is more than one false condition. It saves on multiple indentation